refactored oneClick server route, controller, and model to match our updated data tables and file structure
