+++ server/index.js
@@ -1,4 +1,5 @@
 const employers = require('./controllers/employersController');
+const applications = require('./controllers/applicationsController');
 const express = require('express');
 const app = express();
 const port = 3000;
@@ -22,6 +23,8 @@ app.get('/', (req, res) => {
 // });
 
 app.get('/data/employers', employers.retrieveEmployerData);
+
+app.post('/onclick', applications.oneClickApply);
 
 app.listen(port, () => {
   console.log(`Jobsite app listening at http://localhost:${port}`);
